+++ lib/ftpd/disk_file_system.rb
@@ -377,4 +377,24 @@ module Ftpd
     end
 
   end
+
+  # A disk file system that does not allow any modification (writes,
+  # deletes, etc.)
+
+  class ReadOnlyDiskFileSystem
+
+    include DiskFileSystem::Base
+    include DiskFileSystem::List
+    include DiskFileSystem::Read
+
+    # Make a new instance to serve a directory.  data_dir should be an
+    # absolute path.
+
+    def initialize(data_dir)
+      set_data_dir data_dir
+      translate_exception SystemCallError
+    end
+
+  end
+
 end